♻️ refactor: 공지 내용 스켈레톤 UI 컴포넌트 분리
