+++ src/features/notice/ui/NoticeDetailContent.tsx
@@ -1,6 +1,7 @@
 import { useQuery } from "@tanstack/react-query"
 
 import { getNoticeDetail } from "../api/noticeApi"
+import { NoticeDetailSkeleton } from "./NoticeDetailSkeleton"
 
 interface NoticeDetailContentProps {
   noticeId: string
@@ -65,34 +66,6 @@ export function NoticeDetailContent({ noticeId }: NoticeDetailContentProps) {
             ))}
         </div>
       )}
-    </div>
-  )
-}
-
-export function NoticeDetailSkeleton() {
-  return (
-    <div className="flex animate-pulse flex-col gap-8">
-      <div className="flex flex-col gap-3">
-        <div className="bg-teal-gray-200 h-3.5 w-50 rounded-[4px]"></div>
-        <div className="bg-teal-gray-200 h-3.5 w-full rounded-[4px]"></div>
-        <div className="bg-teal-gray-200 h-3.5 w-full rounded-[4px]"></div>
-        <div className="bg-teal-gray-200 h-3.5 w-full rounded-[4px]"></div>
-        <div className="bg-teal-gray-200 h-3.5 w-full rounded-[4px]"></div>
-      </div>
-      <div className="flex flex-col gap-3">
-        <div className="bg-teal-gray-200 h-3.5 w-50 rounded-[4px]"></div>
-        <div className="bg-teal-gray-200 h-3.5 w-full rounded-[4px]"></div>
-        <div className="bg-teal-gray-200 h-3.5 w-full rounded-[4px]"></div>
-        <div className="bg-teal-gray-200 h-3.5 w-full rounded-[4px]"></div>
-        <div className="bg-teal-gray-200 h-3.5 w-full rounded-[4px]"></div>
-      </div>
-      <div className="flex flex-col gap-3">
-        <div className="bg-teal-gray-200 h-3.5 w-50 rounded-[4px]"></div>
-        <div className="bg-teal-gray-200 h-3.5 w-full rounded-[4px]"></div>
-        <div className="bg-teal-gray-200 h-3.5 w-full rounded-[4px]"></div>
-        <div className="bg-teal-gray-200 h-3.5 w-full rounded-[4px]"></div>
-        <div className="bg-teal-gray-200 h-3.5 w-full rounded-[4px]"></div>
-      </div>
     </div>
   )
 }
