+++ src/features/notice/ui/NoticeDetailSkeleton.tsx
@@ -1,0 +1,27 @@
+export function NoticeDetailSkeleton() {
+  return (
+    <div className="flex animate-pulse flex-col gap-8">
+      <div className="flex flex-col gap-3">
+        <div className="bg-teal-gray-200 h-3.5 w-50 rounded-[4px]"></div>
+        <div className="bg-teal-gray-200 h-3.5 w-full rounded-[4px]"></div>
+        <div className="bg-teal-gray-200 h-3.5 w-full rounded-[4px]"></div>
+        <div className="bg-teal-gray-200 h-3.5 w-full rounded-[4px]"></div>
+        <div className="bg-teal-gray-200 h-3.5 w-full rounded-[4px]"></div>
+      </div>
+      <div className="flex flex-col gap-3">
+        <div className="bg-teal-gray-200 h-3.5 w-50 rounded-[4px]"></div>
+        <div className="bg-teal-gray-200 h-3.5 w-full rounded-[4px]"></div>
+        <div className="bg-teal-gray-200 h-3.5 w-full rounded-[4px]"></div>
+        <div className="bg-teal-gray-200 h-3.5 w-full rounded-[4px]"></div>
+        <div className="bg-teal-gray-200 h-3.5 w-full rounded-[4px]"></div>
+      </div>
+      <div className="flex flex-col gap-3">
+        <div className="bg-teal-gray-200 h-3.5 w-50 rounded-[4px]"></div>
+        <div className="bg-teal-gray-200 h-3.5 w-full rounded-[4px]"></div>
+        <div className="bg-teal-gray-200 h-3.5 w-full rounded-[4px]"></div>
+        <div className="bg-teal-gray-200 h-3.5 w-full rounded-[4px]"></div>
+        <div className="bg-teal-gray-200 h-3.5 w-full rounded-[4px]"></div>
+      </div>
+    </div>
+  )
+}
